PIM-6442: Rename ProductValue factories
