+++ src/Pim/Bundle/CatalogBundle/EventSubscriber/LoadProductValuesSubscriber.php
@@ -5,7 +5,7 @@ namespace Pim\Bundle\CatalogBundle\EventSubscriber;
 use Doctrine\Common\EventSubscriber;
 use Doctrine\Common\Persistence\Event\LifecycleEventArgs;
 use Doctrine\ORM\Events;
-use Pim\Component\Catalog\Factory\ProductValueCollectionFactory;
+use Pim\Component\Catalog\Factory\ValueCollectionFactory;
 use Pim\Component\Catalog\Model\ProductInterface;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
@@ -26,7 +26,7 @@ class LoadProductValuesSubscriber implements EventSubscriber
     /** @var ContainerInterface */
     protected $container;
 
-    /** @var ProductValueCollectionFactory */
+    /** @var ValueCollectionFactory */
     protected $valueCollectionFactory;
 
     /**
@@ -75,12 +75,12 @@ class LoadProductValuesSubscriber implements EventSubscriber
     }
 
     /**
-     * @return ProductValueCollectionFactory
+     * @return ValueCollectionFactory
      */
     private function getProductValueCollectionFactory()
     {
         if (null === $this->valueCollectionFactory) {
-            $this->valueCollectionFactory = $this->container->get('pim_catalog.factory.product_value_collection');
+            $this->valueCollectionFactory = $this->container->get('pim_catalog.factory.value_collection');
         }
 
         return $this->valueCollectionFactory;
