+++ src/Pim/Component/Catalog/Builder/ValuesContainerBuilder.php
@@ -3,7 +3,7 @@
 namespace Pim\Component\Catalog\Builder;
 
 use Pim\Component\Catalog\AttributeTypes;
-use Pim\Component\Catalog\Factory\ProductValueFactory;
+use Pim\Component\Catalog\Factory\ValueFactory;
 use Pim\Component\Catalog\Manager\AttributeValuesResolver;
 use Pim\Component\Catalog\Model\AttributeInterface;
 use Pim\Component\Catalog\Model\ValuesContainerInterface;
@@ -18,12 +18,12 @@ class ValuesContainerBuilder implements ValuesContainerBuilderInterface
     /** @var AttributeValuesResolver */
     protected $valuesResolver;
 
-    /** @var ProductValueFactory */
+    /** @var ValueFactory */
     protected $productValueFactory;
 
     public function __construct(
         AttributeValuesResolver $valuesResolver,
-        ProductValueFactory $productValueFactory
+        ValueFactory $productValueFactory
     ) {
         $this->valuesResolver = $valuesResolver;
         $this->productValueFactory = $productValueFactory;
