+++ src/Pim/Component/Catalog/Completeness/CompletenessCalculator.php
@@ -6,7 +6,7 @@ use Akeneo\Component\StorageUtils\Repository\CachedObjectRepositoryInterface;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Pim\Component\Catalog\Completeness\Checker\ProductValueCompleteCheckerInterface;
-use Pim\Component\Catalog\Factory\ProductValueFactory;
+use Pim\Component\Catalog\Factory\ValueFactory;
 use Pim\Component\Catalog\Model\ChannelInterface;
 use Pim\Component\Catalog\Model\CompletenessInterface;
 use Pim\Component\Catalog\Model\FamilyInterface;
@@ -30,7 +30,7 @@ use Pim\Component\Catalog\Model\ValueCollectionInterface;
  */
 class CompletenessCalculator implements CompletenessCalculatorInterface
 {
-    /** @var ProductValueFactory */
+    /** @var ValueFactory */
     protected $productValueFactory;
 
     /** @var CachedObjectRepositoryInterface */
@@ -46,14 +46,14 @@ class CompletenessCalculator implements CompletenessCalculatorInterface
     protected $completenessClass;
 
     /**
-     * @param ProductValueFactory                  $productValueFactory
+     * @param ValueFactory                         $productValueFactory
      * @param CachedObjectRepositoryInterface      $channelRepository
      * @param CachedObjectRepositoryInterface      $localeRepository
      * @param ProductValueCompleteCheckerInterface $productValueCompleteChecker
      * @param string                               $completenessClass
      */
     public function __construct(
-        ProductValueFactory $productValueFactory,
+        ValueFactory $productValueFactory,
         CachedObjectRepositoryInterface $channelRepository,
         CachedObjectRepositoryInterface $localeRepository,
         ProductValueCompleteCheckerInterface $productValueCompleteChecker,
