+++ src/Pim/Component/Catalog/Factory/ProductValue/MetricValueFactory.php
@@ -9,13 +9,13 @@ use Pim\Component\Catalog\Model\AttributeInterface;
 /**
  * Factory that creates metric product values.
  *
- * @internal  Please, do not use this class directly. You must use \Pim\Component\Catalog\Factory\ProductValueFactory.
+ * @internal  Please, do not use this class directly. You must use \Pim\Component\Catalog\Factory\ValueFactory.
  *
  * @author    Damien Carcel (damien.carcel@akeneo.com)
  * @copyright 2017 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  */
-class MetricProductValueFactory implements ProductValueFactoryInterface
+class MetricValueFactory implements ProductValueFactoryInterface
 {
     /** @var MetricFactory */
     protected $metricFactory;
