--- src/Pim/Component/Catalog/Factory/ProductValue/OptionProductValueFactory.php
@@ -1,124 +1,0 @@
-<?php
-
-namespace Pim\Component\Catalog\Factory\ProductValue;
-
-use Akeneo\Component\StorageUtils\Exception\InvalidPropertyException;
-use Akeneo\Component\StorageUtils\Exception\InvalidPropertyTypeException;
-use Pim\Component\Catalog\Exception\InvalidOptionException;
-use Pim\Component\Catalog\Model\AttributeInterface;
-use Pim\Component\Catalog\Model\AttributeOptionInterface;
-use Pim\Component\Catalog\Repository\AttributeOptionRepositoryInterface;
-
-/**
- * Factory that creates option (simple-select) product values.
- *
- * @internal  Please, do not use this class directly. You must use \Pim\Component\Catalog\Factory\ProductValueFactory.
- *
- * @author    Damien Carcel (damien.carcel@akeneo.com)
- * @copyright 2017 Akeneo SAS (http://www.akeneo.com)
- * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
- */
-class OptionProductValueFactory implements ProductValueFactoryInterface
-{
-    /** @var AttributeOptionRepositoryInterface */
-    protected $attrOptionRepository;
-
-    /** @var string */
-    protected $productValueClass;
-
-    /** @var string */
-    protected $supportedAttributeType;
-
-    /**
-     * @param AttributeOptionRepositoryInterface $attrOptionRepository
-     * @param string                             $productValueClass
-     * @param string                             $supportedAttributeType
-     */
-    public function __construct(
-        AttributeOptionRepositoryInterface $attrOptionRepository,
-        $productValueClass,
-        $supportedAttributeType
-    ) {
-        $this->attrOptionRepository = $attrOptionRepository;
-        $this->productValueClass = $productValueClass;
-        $this->supportedAttributeType = $supportedAttributeType;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function create(AttributeInterface $attribute, $channelCode, $localeCode, $data)
-    {
-        $this->checkData($attribute, $data);
-
-        if (null !== $data) {
-            $data = $this->getOption($attribute, $data);
-        }
-
-        $value = new $this->productValueClass($attribute, $channelCode, $localeCode, $data);
-
-        return $value;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function supports($attributeType)
-    {
-        return $attributeType === $this->supportedAttributeType;
-    }
-
-    /**
-     * Checks if data is valid.
-     *
-     * @param AttributeInterface $attribute
-     * @param mixed              $data
-     *
-     * @throws InvalidPropertyTypeException
-     */
-    protected function checkData(AttributeInterface $attribute, $data)
-    {
-        if (null === $data) {
-            return;
-        }
-
-        if (!is_string($data) && !is_numeric($data)) {
-            throw InvalidPropertyTypeException::stringExpected(
-                $attribute->getCode(),
-                static::class,
-                $data
-            );
-        }
-    }
-
-    /**
-     * Gets an attribute option from its code.
-     *
-     * @param AttributeInterface $attribute
-     * @param string|null        $optionCode
-     *
-     * @throws InvalidOptionException
-     * @return AttributeOptionInterface|null
-     */
-    protected function getOption(AttributeInterface $attribute, $optionCode)
-    {
-        if (null === $optionCode) {
-            return null;
-        }
-
-        $identifier = $attribute->getCode() . '.' . $optionCode;
-        $option = $this->attrOptionRepository->findOneByIdentifier($identifier);
-
-        if (null === $option) {
-            throw InvalidOptionException::validEntityCodeExpected(
-                $attribute->getCode(),
-                'code',
-                'The option does not exist',
-                static::class,
-                $optionCode
-            );
-        }
-
-        return $option;
-    }
-}
