+++ src/Pim/Component/Catalog/Factory/ProductValue/OptionsValueFactory.php
@@ -11,13 +11,13 @@ use Pim\Component\Catalog\Repository\AttributeOptionRepositoryInterface;
 /**
  * Factory that creates options (multi-select) product values.
  *
- * @internal  Please, do not use this class directly. You must use \Pim\Component\Catalog\Factory\ProductValueFactory.
+ * @internal  Please, do not use this class directly. You must use \Pim\Component\Catalog\Factory\ValueFactory.
  *
  * @author    Damien Carcel (damien.carcel@akeneo.com)
  * @copyright 2017 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  */
-class OptionsProductValueFactory implements ProductValueFactoryInterface
+class OptionsValueFactory implements ProductValueFactoryInterface
 {
     /** @var AttributeOptionRepositoryInterface */
     protected $attrOptionRepository;
