+++ src/Pim/Component/Catalog/Factory/ProductValue/PriceCollectionValueFactory.php
@@ -11,13 +11,13 @@ use Pim\Component\Catalog\Model\PriceCollection;
 /**
  * Factory that creates price collection product values.
  *
- * @internal  Please, do not use this class directly. You must use \Pim\Component\Catalog\Factory\ProductValueFactory.
+ * @internal  Please, do not use this class directly. You must use \Pim\Component\Catalog\Factory\ValueFactory.
  *
  * @author    Damien Carcel (damien.carcel@akeneo.com)
  * @copyright 2017 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  */
-class PriceCollectionProductValueFactory implements ProductValueFactoryInterface
+class PriceCollectionValueFactory implements ProductValueFactoryInterface
 {
     /** @var PriceFactory */
     protected $priceFactory;
