+++ src/Pim/Component/Catalog/Factory/ProductValue/ScalarValueFactory.php
@@ -9,13 +9,13 @@ use Pim\Component\Catalog\Model\AttributeInterface;
 /**
  * Factory that creates simple product values (text, textarea and number).
  *
- * @internal  Please, do not use this class directly. You must use \Pim\Component\Catalog\Factory\ProductValueFactory.
+ * @internal  Please, do not use this class directly. You must use \Pim\Component\Catalog\Factory\ValueFactory.
  *
  * @author    Samir Boulil <samir.boulil@akeneo.com>
  * @copyright 2016 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class ScalarProductValueFactory implements ProductValueFactoryInterface
+class ScalarValueFactory implements ProductValueFactoryInterface
 {
     /** @var string */
     protected $productValueClass;
