+++ src/Pim/Component/Catalog/Factory/ValueFactory.php
@@ -12,13 +12,13 @@ use Pim\Component\Catalog\Validator\AttributeValidatorHelper;
  * Factory that creates product values.
  *
  * "RegisterProductValueFactoryPass" allows to register private product value
- * factories tagged with "pim_catalog.factory.product_value".
+ * factories tagged with "pim_catalog.factory.value".
  *
  * @author    Samir Boulil <samir.boulil@akeneo.com>
  * @copyright 2016 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class ProductValueFactory
+class ValueFactory
 {
     /** @var AttributeValidatorHelper */
     protected $attributeValidatorHelper;
