+++ src/Pim/Component/Catalog/Manager/ProductTemplateMediaManager.php
@@ -3,7 +3,7 @@
 namespace Pim\Component\Catalog\Manager;
 
 use Akeneo\Component\FileStorage\File\FileStorerInterface;
-use Pim\Component\Catalog\Factory\ProductValueFactory;
+use Pim\Component\Catalog\Factory\ValueFactory;
 use Pim\Component\Catalog\FileStorage;
 use Pim\Component\Catalog\Model\ProductTemplateInterface;
 use Pim\Component\Catalog\ProductValue\MediaValueInterface;
@@ -20,16 +20,16 @@ class ProductTemplateMediaManager
     /** @var FileStorerInterface */
     protected $fileStorer;
 
-    /** @var ProductValueFactory */
+    /** @var ValueFactory */
     protected $productValueFactory;
 
     /**
      * @param FileStorerInterface $fileStorer
-     * @param ProductValueFactory $productValueFactory
+     * @param ValueFactory        $productValueFactory
      */
     public function __construct(
         FileStorerInterface $fileStorer,
-        ProductValueFactory $productValueFactory
+        ValueFactory $productValueFactory
     ) {
         $this->fileStorer = $fileStorer;
         $this->productValueFactory = $productValueFactory;
